Fix smart_charging()
Fix smart_charging() method and add bool to turn on or off. Original API call to /api/chargers/{id}/commands/smart_charging was believed to be a toggle, but updated docs says it is only to turn on and that settings must be used to turn off: «Turn off smart charging again by posting an explicit value to the charger settings endpoint».

So to summarize, why not use call to /api/chargers/{id}/settings to turn both on and off. Tested to be working with HA switch as well.